+++ android/src/main/java/com/mrousavy/camera/extensions/Context+displayRotation.kt
@@ -1,4 +1,4 @@
-package com.mrousavy.camera.utils
+package com.mrousavy.camera.extensions
 
 import android.content.Context
 import android.os.Build
