+++ android/src/main/java/com/mrousavy/camera/extensions/List+containsAny.kt
@@ -1,4 +1,4 @@
-package com.mrousavy.camera.utils
+package com.mrousavy.camera.extensions
 
 fun <T> List<T>.containsAny(elements: List<T>): Boolean {
   return elements.any { element -> this.contains(element) }
