+++ android/src/main/java/com/mrousavy/camera/extensions/Size+Extensions.kt
@@ -1,4 +1,4 @@
-package com.mrousavy.camera.utils
+package com.mrousavy.camera.extensions
 
 import android.util.Size
 import android.util.SizeF
