+++ android/src/main/java/com/mrousavy/camera/extensions/WritableArray+Nullables.kt
@@ -1,4 +1,4 @@
-package com.mrousavy.camera.utils
+package com.mrousavy.camera.extensions
 
 import com.facebook.react.bridge.WritableArray
